+++ test.js
@@ -1,0 +1,3 @@
+let esiJS = require('./endpoints') 
+
+console.log(esiJS.util.getSettings())